Commit #21.1

+++ Produzione/coda.c
@@ -1,7 +1,7 @@
 #include "coda.h"
 #include "newHash.h"
 
-const int dimensioneTotale = 1;
+const int dimensioneTotale = 500000;
 int limite_dimensione = dimensioneTotale;
 
 struct nodo* coda; //puntatore alla coda della lista collegata
@@ -17,12 +17,14 @@ struct nodo* libera_n(struct nodo *testa, struct nodo * n){
       
       // Capire perché non funziona il pthread_rwlock_init()
 
-      /*       if(pthread_rwlock_trywrlock(n->hashItem->sem)!=0){ 
-        printf("Errno number is: %d", errno);
-            //if fail->
-            return NULL;
-      } */
-      
+    if(pthread_rwlock_trywrlock(n->hashItem->sem)!=0){ 
+          printf("Errno number is: %d", errno);
+              //if fail->
+              return NULL;
+        }
+/*       printf("\nBefore deadlock\n");
+      pthread_rwlock_wrlock(n->hashItem->sem);
+      printf("\nAfter deadlock\n"); */
       //se fail->return(puo' capitare che anche se era il nodo meno usato ci sta qualcuno che lo sta leggendo? non dovrebbe capitare. Anche se capitasse ritornando verrebbe cercato un nuovo nodo da eliminare(eventualmente di nuovo lui)) :
       //se va bene blocco il semaforo in scrittura relativo a testa->sem(non c' e' bisogno di liberarlo perche' elimino il nodo) 
       limite_dimensione+=deleteHashNode(n->indice);//CONTROLLARE perche' la funzione puo' ritornare NULL
@@ -72,7 +74,6 @@ struct nodo* inserisci_in_coda(struct nodo *coda,struct nodo *mes){//si potrebbe
     coda=mes; //aggiorno la coda
     mes->suc=NULL; //il puntatore (della coda) al nuovo nodo deve essere NULL (nel caso ho reinserimento di un nodo precedente se non metto questa riga avro' un ciclo infinito in quanto il nodo reinserito in passato puntava ad un altro nodo)
   }
-  printf("Sem address in insert in coda: %x", mes->hashItem->sem);
   return coda;
 }
 
@@ -126,7 +127,6 @@ struct nodo* inserisci_n(struct nodo* testa,long nod,struct hashNode *h){
     n2->indice=nod;
     n2->suc=NULL;
     n2->hashItem=h;
-    printf("Sem address in insert: %x", n2->hashItem->sem);
   coda=inserisci_in_coda(coda,n2);  
   testa=aus2; //aus2 tiene conto di tutte le modifiche eventuali della testa della lista
   return testa;
